[CHANGE] Combat is implemented, basics server & client side. [CHANGE] Rendering is overhauled, it now works based on a zOrdering system, this is the base now that everything is built upon. [ADD] Hitsplats are now in game. [ADD] Healthbars are now implemented. [FIX] The goblin its offset is moreso corrected.

+++ Client/include/Core/Rendering/Sprite/Sprite.cpp
@@ -9,7 +9,7 @@ const Graphics::SpriteType Graphics::_Sprite::get_sprite_type() const
 	return m_eSpriteType;
 }
 
-const uint32_t Graphics::_Sprite::get_framecount() const
+const uint8_t Graphics::_Sprite::get_framecount() const
 {
 	return m_frameCount;
 }
